Modified the utils directory architecture
